+++ src/components/CardNumberInput/constants/index.js
@@ -1,0 +1,4 @@
+import VALIDATION_REGEXES from './validation-regexes';
+
+export * from './card-schemes';
+export { VALIDATION_REGEXES };
