+++ src/components/CardNumberInput/index.js
@@ -1,4 +1,7 @@
 import CardNumberInput from './CardNumberInput';
 
 export * from './components';
+export { SCHEMES } from './constants';
+export { CardNumberInput };
+
 export default CardNumberInput;
